Feat: add top reducer file for import per reducer
